Upgrade to gulp@^4
