+++ package.json
@@ -9,21 +9,21 @@
     "build": "gulp build",
     "clean": "gulp clean",
     "lint": "eslint ./src/ ./test/ ./vendor/ gulpfile.babel.js",
-    "make:darwin": "gulp make:darwin",
-    "make:linux": "gulp make:linux",
-    "make:deb": "gulp make:linux:deb",
-    "make:rpm": "gulp make:linux:rpm",
-    "make:deb:32": "gulp deb:linux:32",
-    "make:deb:64": "gulp deb:linux:64",
-    "make:rpm:32": "gulp rpm:linux:32",
-    "make:rpm:64": "gulp rpm:linux:64",
-    "make:win": "gulp make:win",
-    "make:win:uwp": "gulp make:win:uwp",
-    "package:darwin": "gulp package:darwin",
-    "package:linux": "gulp package:linux",
-    "package:linux:32": "gulp package:linux:32",
-    "package:linux:64": "gulp package:linux:64",
-    "package:win": "gulp package:win",
+    "make:darwin": "gulp makeDarwin",
+    "make:linux": "gulp makeLinux",
+    "make:deb": "gulp makeLinuxDeb",
+    "make:rpm": "gulp makeLinuxRpm",
+    "make:deb:32": "gulp debLinux32",
+    "make:deb:64": "gulp debLinux64",
+    "make:rpm:32": "gulp rpmLinux32",
+    "make:rpm:64": "gulp rpmLinux64",
+    "make:win": "gulp makeWin",
+    "make:win:uwp": "gulp makeWinUwp",
+    "package:darwin": "gulp packageDarwin",
+    "package:linux": "gulp packageLinux",
+    "package:linux:32": "gulp packageLinux32",
+    "package:linux:64": "gulp packageLinux64",
+    "package:win": "gulp packageWin",
     "postinstall": "electron-rebuild --types=prod,optional",
     "postversion": "node vendor/postversion.js",
     "test": "npm run lint",
@@ -61,6 +61,7 @@
     "@slack/client": "^4.8.0",
     "archiver": "^1.0.1",
     "auto-launch": "^4.0.0",
+    "del": "^5.1.0",
     "discord-rich-presence": "^0.0.7",
     "electron-chromecast": "^1.1.0",
     "gmusic-mini-player.js": "^2.0.10",
@@ -71,7 +72,7 @@
     "jquery": "^2.2.4",
     "lastfm": "^0.9.2",
     "listenbrainz": "^1.0.4",
-    "lodash": "^4.17.13",
+    "lodash": "^4.17.15",
     "material-ui": "^0.15.4",
     "mdns": "^2.3.3",
     "mkdirp": "^0.5.1",
@@ -118,17 +119,15 @@
     "eslint-plugin-jsx-a11y": "^1.5.5",
     "eslint-plugin-react": "^5.2.2",
     "fs-extra": "^0.30.0",
-    "glob": "^7.0.6",
+    "glob": "^7.1.6",
     "gpmdp-electron-winstaller": "1.0.0",
-    "gulp": "^3.9.1",
+    "gulp": "^4.0.2",
     "gulp-babel": "^6.1.2",
-    "gulp-clean": "^0.3.2",
     "gulp-concat": "^2.6.0",
-    "gulp-cssmin": "^0.1.7",
-    "gulp-header": "^1.8.2",
-    "gulp-less": "^3.1.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-uglify": "^2.0.0",
+    "gulp-cssmin": "^0.2.0",
+    "gulp-header": "^2.0.9",
+    "gulp-less": "^4.0.1",
+    "gulp-replace": "^1.0.0",
     "istanbul": "^1.1.0-alpha.1",
     "jimp": "^0.2.27",
     "material-design-icons-iconfont": "^2.0.5",
@@ -145,7 +144,6 @@
     "rcedit": "^0.7.0",
     "react-addons-test-utils": "^15.3.1",
     "remap-istanbul": "0.6.4",
-    "run-sequence": "^1.1.5",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
     "source-map": "0.5.6",
